epic yak shave
turned out magnum had some python deps that were newer than other stuff
being installed which meant there were some newer features not there
that it needed to do a db-migrate etc.

So I decided to put it in its own virtualenv and quickly realized
that the best way to do that would be to put all of them into a
virtualenv that resembles how the giftwrap packages get laid down and used.

This change means that along with magnum working ( which we can feature
flag off if we don't want to install it in CI etc as it's probably
still 'in progress' as far as supportability goes ) we know have a
source install that ends up looking almost identical to a package
install.

It also means that we're declaring the pip and system dependencies
for each project in its source install which means that we should
have everything required to create the giftwrap manifest for a
release in the ursula vars making it much easier to prepare packages.